Bring resource creation, patching and deletion into a single class hierarchy
