+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/operations/ResourceOperation.java
@@ -22,6 +22,8 @@ import io.fabric8.kubernetes.api.model.ConfigMapList;
 import io.fabric8.kubernetes.api.model.DoneableConfigMap;
 import io.fabric8.kubernetes.api.model.DoneablePersistentVolumeClaim;
 import io.fabric8.kubernetes.api.model.DoneableService;
+import io.fabric8.kubernetes.api.model.HasMetadata;
+import io.fabric8.kubernetes.api.model.KubernetesResource;
 import io.fabric8.kubernetes.api.model.PersistentVolumeClaim;
 import io.fabric8.kubernetes.api.model.PersistentVolumeClaimList;
 import io.fabric8.kubernetes.api.model.Service;
@@ -34,6 +36,7 @@ import io.fabric8.kubernetes.api.model.extensions.StatefulSet;
 import io.fabric8.kubernetes.api.model.extensions.StatefulSetList;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.dsl.MixedOperation;
+import io.fabric8.kubernetes.client.dsl.Patchable;
 import io.fabric8.kubernetes.client.dsl.Resource;
 import io.fabric8.kubernetes.client.dsl.RollableScalableResource;
 import io.fabric8.kubernetes.client.dsl.ScalableResource;
@@ -54,7 +57,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * Abstract resource deletion.
+ * Abstract resource creation, for a generic resource type {@code R}.
  * This class applies the template method pattern, first checking whether the resource exists,
  * and creating it if it does not. It is not an error if the resource did already exist.
  * @param <C> The type of client used to interact with kubernetes.
@@ -63,21 +66,52 @@ import org.slf4j.LoggerFactory;
  * @param <D> The doneable variant of the Kubernetes resource type.
  * @param <R> The resource operations.
  */
-public abstract class DeleteOperation<C, T, L, D, R extends Resource<T, D>> {
+public abstract class ResourceOperation<C, T extends HasMetadata, L, D, R extends Resource<T, D>> {
 
-    private static final Logger log = LoggerFactory.getLogger(DeleteOperation.class);
-
-    private final String resourceKind;
+    private static final Logger log = LoggerFactory.getLogger(ResourceOperation.class);
     private final Vertx vertx;
     private final C client;
+    private final String resourceKind;
 
-    public DeleteOperation(Vertx vertx, C client, String resourceKind) {
+    public ResourceOperation(Vertx vertx, C client, String resourceKind) {
         this.vertx = vertx;
         this.client = client;
         this.resourceKind = resourceKind;
     }
 
     protected abstract MixedOperation<T, L, D, R> operation();
+
+    public void create(T resource, Handler<AsyncResult<Void>> handler) {
+        vertx.createSharedWorkerExecutor("kubernetes-ops-pool").executeBlocking(
+                future -> {
+                    if (operation().inNamespace(resource.getMetadata().getNamespace()).withName(resource.getMetadata().getName()).get() == null) {
+                        try {
+                            log.info("Creating {} {}", resourceKind, resource);
+                            operation().createOrReplace(resource);
+                            future.complete();
+                        } catch (Exception e) {
+                            log.error("Caught exception while creating {}", resourceKind, e);
+                            future.fail(e);
+                        }
+                    }
+                    else {
+                        log.warn("{} {} already exists", resourceKind, resource);
+                        future.complete();
+                    }
+                },
+                false,
+                res -> {
+                    if (res.succeeded()) {
+                        log.info("{} {} has been created", resourceKind, resource);
+                        handler.handle(Future.succeededFuture());
+                    }
+                    else {
+                        log.error("{} creation failed:", resourceKind, res.cause());
+                        handler.handle(Future.failedFuture(res.cause()));
+                    }
+                }
+        );
+    }
 
     public void delete(String namespace, String name, Handler<AsyncResult<Void>> handler) {
         vertx.createSharedWorkerExecutor("kubernetes-ops-pool").executeBlocking(
@@ -110,9 +144,36 @@ public abstract class DeleteOperation<C, T, L, D, R extends Resource<T, D>> {
         );
     }
 
-    public static DeleteOperation<KubernetesClient, Deployment, DeploymentList, DoneableDeployment, ScalableResource<Deployment, DoneableDeployment>> deleteDeployment(Vertx vertx, KubernetesClient client) {
-        return new DeleteOperation<KubernetesClient, Deployment, DeploymentList, DoneableDeployment, ScalableResource<Deployment, DoneableDeployment>>(vertx, client, "Deployment") {
+    public <P extends Patchable<T, T>> void patch(P patchable, T patch, Handler<AsyncResult<Void>> handler) {
+        vertx.createSharedWorkerExecutor("kubernetes-ops-pool").executeBlocking(
+                future -> {
+                    try {
+                        log.info("Patching resource with {}", patch);
+                        patchable.patch(patch);
+                        future.complete();
+                    }
+                    catch (Exception e) {
+                        log.error("Caught exception while patching", e);
+                        future.fail(e);
+                    }
+                },
+                false,
+                res -> {
+                    if (res.succeeded()) {
+                        log.info("Resource has been patched", patch);
+                        handler.handle(Future.succeededFuture());
+                    }
+                    else {
+                        log.error("Failed to patch resource: {}", res.cause().toString());
+                        handler.handle(Future.failedFuture(res.cause()));
+                    }
+                }
+        );
+    }
 
+
+    public static ResourceOperation<KubernetesClient, Deployment, DeploymentList, DoneableDeployment, ScalableResource<Deployment, DoneableDeployment>> deployment(Vertx vertx, KubernetesClient client) {
+        return new ResourceOperation<KubernetesClient, Deployment, DeploymentList, DoneableDeployment, ScalableResource<Deployment, DoneableDeployment>>(vertx, client, "Deployment") {
             @Override
             protected MixedOperation<Deployment, DeploymentList, DoneableDeployment, ScalableResource<Deployment, DoneableDeployment>> operation() {
                 return client.extensions().deployments();
@@ -120,8 +181,8 @@ public abstract class DeleteOperation<C, T, L, D, R extends Resource<T, D>> {
         };
     }
 
-    public static DeleteOperation<KubernetesClient, ConfigMap, ConfigMapList, DoneableConfigMap, Resource<ConfigMap, DoneableConfigMap>> deleteConfigMap(Vertx vertx, KubernetesClient client) {
-        return new DeleteOperation<KubernetesClient, ConfigMap, ConfigMapList, DoneableConfigMap, Resource<ConfigMap, DoneableConfigMap>>(vertx, client, "ConfigMap") {
+    public static ResourceOperation<KubernetesClient, ConfigMap, ConfigMapList, DoneableConfigMap, Resource<ConfigMap, DoneableConfigMap>> configMap(Vertx vertx, KubernetesClient client) {
+        return new ResourceOperation<KubernetesClient, ConfigMap, ConfigMapList, DoneableConfigMap, Resource<ConfigMap, DoneableConfigMap>>(vertx, client, "ConfigMap") {
 
             @Override
             protected MixedOperation<ConfigMap, ConfigMapList, DoneableConfigMap, Resource<ConfigMap, DoneableConfigMap>> operation() {
@@ -130,17 +191,20 @@ public abstract class DeleteOperation<C, T, L, D, R extends Resource<T, D>> {
         };
     }
 
-    public static DeleteOperation<KubernetesClient, StatefulSet, StatefulSetList, DoneableStatefulSet, RollableScalableResource<StatefulSet, DoneableStatefulSet>> deleteStatefulSet(Vertx vertx, KubernetesClient client) {
-        return new DeleteOperation<KubernetesClient, StatefulSet, StatefulSetList, DoneableStatefulSet, RollableScalableResource<StatefulSet, DoneableStatefulSet>>(vertx, client, "StatefulSet") {
+    public static ResourceOperation<KubernetesClient, StatefulSet, StatefulSetList, DoneableStatefulSet, RollableScalableResource<StatefulSet, DoneableStatefulSet>> statefulSet(Vertx vertx, KubernetesClient client) {
+        return new ResourceOperation<KubernetesClient, StatefulSet, StatefulSetList, DoneableStatefulSet, RollableScalableResource<StatefulSet, DoneableStatefulSet>>(vertx, client, "StatefulSet") {
+
             @Override
             protected MixedOperation<StatefulSet, StatefulSetList, DoneableStatefulSet, RollableScalableResource<StatefulSet, DoneableStatefulSet>> operation() {
                 return client.apps().statefulSets();
             }
+
         };
     }
 
-    public static DeleteOperation<KubernetesClient, Service, ServiceList, DoneableService, Resource<Service, DoneableService>> deleteService(Vertx vertx, KubernetesClient client) {
-        return new DeleteOperation<KubernetesClient, Service, ServiceList, DoneableService, Resource<Service, DoneableService>>(vertx, client, "Service") {
+    public static ResourceOperation<KubernetesClient, Service, ServiceList, DoneableService, Resource<Service, DoneableService>> service(Vertx vertx, KubernetesClient client) {
+        return new ResourceOperation<KubernetesClient, Service, ServiceList, DoneableService, Resource<Service, DoneableService>>(vertx, client, "Service") {
+
             @Override
             protected MixedOperation<Service, ServiceList, DoneableService, Resource<Service, DoneableService>> operation() {
                 return client.services();
@@ -148,8 +212,8 @@ public abstract class DeleteOperation<C, T, L, D, R extends Resource<T, D>> {
         };
     }
 
-    public static DeleteOperation<OpenShiftClient, BuildConfig, BuildConfigList, DoneableBuildConfig, BuildConfigResource<BuildConfig, DoneableBuildConfig, Void, Build>> deleteBuildConfig(Vertx vertx, OpenShiftClient client) {
-        return new DeleteOperation<OpenShiftClient, BuildConfig, BuildConfigList, DoneableBuildConfig, BuildConfigResource<BuildConfig, DoneableBuildConfig, Void, Build>>(vertx, client, "BuildConfig") {
+    public static ResourceOperation<OpenShiftClient, BuildConfig, BuildConfigList, DoneableBuildConfig, BuildConfigResource<BuildConfig, DoneableBuildConfig, Void, Build>> buildConfig(Vertx vertx, OpenShiftClient client) {
+        return new ResourceOperation<OpenShiftClient, BuildConfig, BuildConfigList, DoneableBuildConfig, BuildConfigResource<BuildConfig, DoneableBuildConfig, Void, Build>>(vertx, client, "BuildConfig") {
             @Override
             protected MixedOperation<BuildConfig, BuildConfigList, DoneableBuildConfig, BuildConfigResource<BuildConfig, DoneableBuildConfig, Void, Build>> operation() {
                 return client.buildConfigs();
@@ -157,8 +221,8 @@ public abstract class DeleteOperation<C, T, L, D, R extends Resource<T, D>> {
         };
     }
 
-    public static DeleteOperation<OpenShiftClient, ImageStream, ImageStreamList, DoneableImageStream, Resource<ImageStream, DoneableImageStream>> deleteImageStream(Vertx vertx, OpenShiftClient client) {
-        return new DeleteOperation<OpenShiftClient, ImageStream, ImageStreamList, DoneableImageStream, Resource<ImageStream, DoneableImageStream>>(vertx, client, "ImageStream") {
+    public static ResourceOperation<OpenShiftClient, ImageStream, ImageStreamList, DoneableImageStream, Resource<ImageStream, DoneableImageStream>> imageStream(Vertx vertx, OpenShiftClient client) {
+        return new ResourceOperation<OpenShiftClient, ImageStream, ImageStreamList, DoneableImageStream, Resource<ImageStream, DoneableImageStream>>(vertx, client, "ImageStream") {
             @Override
             protected MixedOperation<ImageStream, ImageStreamList, DoneableImageStream, Resource<ImageStream, DoneableImageStream>> operation() {
                 return client.imageStreams();
@@ -166,11 +230,19 @@ public abstract class DeleteOperation<C, T, L, D, R extends Resource<T, D>> {
         };
     }
 
-    public static DeleteOperation<KubernetesClient, PersistentVolumeClaim, PersistentVolumeClaimList, DoneablePersistentVolumeClaim, Resource<PersistentVolumeClaim, DoneablePersistentVolumeClaim>> deletePersistentVolumeClaim(Vertx vertx, KubernetesClient client) {
-        return new DeleteOperation<KubernetesClient, PersistentVolumeClaim, PersistentVolumeClaimList, DoneablePersistentVolumeClaim, Resource<PersistentVolumeClaim, DoneablePersistentVolumeClaim>>(vertx, client, "PersistentVolumeClaim") {
+    public static ResourceOperation<KubernetesClient, PersistentVolumeClaim, PersistentVolumeClaimList, DoneablePersistentVolumeClaim, Resource<PersistentVolumeClaim, DoneablePersistentVolumeClaim>> persistentVolumeClaim(Vertx vertx, KubernetesClient client) {
+        return new ResourceOperation<KubernetesClient, PersistentVolumeClaim, PersistentVolumeClaimList, DoneablePersistentVolumeClaim, Resource<PersistentVolumeClaim, DoneablePersistentVolumeClaim>>(vertx, client, "PersistentVolumeClaim") {
             @Override
             protected MixedOperation<PersistentVolumeClaim, PersistentVolumeClaimList, DoneablePersistentVolumeClaim, Resource<PersistentVolumeClaim, DoneablePersistentVolumeClaim>> operation() {
                 return client.persistentVolumeClaims();
+            }
+            @Override
+            public void create(PersistentVolumeClaim resource, Handler<AsyncResult<Void>> handler) {
+                throw new UnsupportedOperationException();// should never happen
+            }
+            @Override
+            public <P extends Patchable<PersistentVolumeClaim, PersistentVolumeClaim>> void patch(P patchable, PersistentVolumeClaim patch, Handler<AsyncResult<Void>> handler) {
+                throw new UnsupportedOperationException();// should never happen
             }
         };
     }
