+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/operations/openshift/UpdateS2IOperation.java
@@ -1,7 +1,16 @@
 package io.strimzi.controller.cluster.operations.openshift;
 
+import io.fabric8.kubernetes.client.dsl.Resource;
+import io.fabric8.openshift.api.model.Build;
+import io.fabric8.openshift.api.model.BuildConfig;
+import io.fabric8.openshift.api.model.BuildConfigList;
+import io.fabric8.openshift.api.model.DoneableBuildConfig;
+import io.fabric8.openshift.api.model.DoneableImageStream;
+import io.fabric8.openshift.api.model.ImageStream;
+import io.fabric8.openshift.api.model.ImageStreamList;
 import io.fabric8.openshift.client.OpenShiftClient;
-import io.strimzi.controller.cluster.operations.kubernetes.PatchOperation;
+import io.fabric8.openshift.client.dsl.BuildConfigResource;
+import io.strimzi.controller.cluster.operations.ResourceOperation;
 import io.strimzi.controller.cluster.resources.Source2Image;
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
@@ -18,8 +27,8 @@ import java.util.List;
 public class UpdateS2IOperation extends S2IOperation {
 
     private static final Logger log = LoggerFactory.getLogger(UpdateS2IOperation.class.getName());
-
-    private final PatchOperation patchOperation;
+    private final ResourceOperation<OpenShiftClient, BuildConfig, BuildConfigList, DoneableBuildConfig, BuildConfigResource<BuildConfig, DoneableBuildConfig, Void, Build>> buildConfigOperation;
+    private final ResourceOperation<OpenShiftClient, ImageStream, ImageStreamList, DoneableImageStream, Resource<ImageStream, DoneableImageStream>> buildImageStreamOperation;
 
     /**
      * Constructor
@@ -28,7 +37,8 @@ public class UpdateS2IOperation extends S2IOperation {
      */
     public UpdateS2IOperation(Vertx vertx, OpenShiftClient client, Source2Image s2i) {
         super(vertx, client,"update", s2i);
-        patchOperation = new PatchOperation(vertx);
+        buildConfigOperation = ResourceOperation.buildConfig(vertx, client);
+        buildImageStreamOperation = ResourceOperation.imageStream(vertx, client);
     }
 
 
@@ -38,7 +48,7 @@ public class UpdateS2IOperation extends S2IOperation {
             List<Future> result = new ArrayList<>(3);
             Future<Void> futureSourceImageStream = Future.future();
 
-            patchOperation.patch(
+            buildImageStreamOperation.patch(
                     client.imageStreams().inNamespace(s2i.getNamespace()).withName(s2i.getSourceImageStreamName()),
                     s2i.patchSourceImageStream(
                             client.imageStreams().inNamespace(s2i.getNamespace()).withName(s2i.getSourceImageStreamName()).get()),
@@ -46,7 +56,7 @@ public class UpdateS2IOperation extends S2IOperation {
             result.add(futureSourceImageStream);
 
             Future<Void> futureTargetImageStream = Future.future();
-            patchOperation
+            buildImageStreamOperation
                     .patch(client.imageStreams().inNamespace(s2i.getNamespace()).withName(s2i.getName()),
                             s2i.patchTargetImageStream(
                                     client.imageStreams().inNamespace(s2i.getNamespace()).withName(s2i.getName()).get()),
@@ -54,7 +64,7 @@ public class UpdateS2IOperation extends S2IOperation {
             result.add(futureTargetImageStream);
 
             Future<Void> futureBuildConfig = Future.future();
-            patchOperation
+            buildConfigOperation
                     .patch(client.buildConfigs().inNamespace(s2i.getNamespace()).withName(s2i.getName()),
                             s2i.patchBuildConfig(
                                     client.buildConfigs().inNamespace(s2i.getNamespace()).withName(s2i.getName()).get()),
